requirements.txt
langchain
langchain-community
langchain-core
sentence-transformers
faiss-cpu
huggingface-hub
python-dotenv
langchain_huggingface
streamlit
pickle
gdown